% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRankedGenes.R
\name{exportRankedGenes}
\alias{exportRankedGenes}
\title{Export Ranked Genes with Cox Hazard Ratios to a CSV file}
\usage{
exportRankedGenes(lhr_results, file_path = "ranked_genes.csv")
}
\arguments{
\item{lhr_results}{A named numeric vector (output from getLHR) where names are gene symbols and values are log hazard ratios.}

\item{file_path}{A character string specifying the file path to save the results. Default is "Ranked_Genes.csv".}
}
\value{
The file path where the ranked gene list is saved.
}
\description{
Saves a ranked list of genes with their log hazard ratios to a specified file.
}
