% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichPlot.R
\name{getEnrichPlot}
\alias{getEnrichPlot}
\title{Enrichment plot for an individual pathway}
\usage{
getEnrichPlot(pathways, pathwayName, stats)
}
\arguments{
\item{pathways}{the human pathways imported from Reactome. Same object as input in getSGSEA function.}

\item{pathwayName}{enter the whole name of a specific pathway from your result table with the quotation mark.}

\item{stats}{log hazard ratios or log fold changes.}
}
\value{
the enrichment plot for an individual pathway
}
\description{
Enrichment plot for an individual pathway
}
