% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGO.R
\name{getGO}
\alias{getGO}
\title{Gets gene sets with Hugo gene symbols to use with SGSEA analysis based on GO terms.}
\usage{
getGO(species = "human")
}
\arguments{
\item{species}{currently accepts 'human' or 'mouse'}
}
\value{
A list of gene sets from the Gene Ontology database
}
\description{
Gets gene sets with Hugo gene symbols to use with SGSEA analysis based on GO terms.
}
