% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLFC.R
\name{getLFC}
\alias{getLFC}
\title{Generate log fold change from the negative binomial generalized linear model}
\usage{
getLFC(countData, label1, label2, numRep)
}
\arguments{
\item{countData}{a data frame contains only counts. Rows should be sample IDs and columns are the gene symbols and should be unique. Data should be filtered before entering this step.}

\item{label1}{group label for the samples.}

\item{label2}{group label for the samples.}

\item{numRep}{Number of replications for each condition.}
}
\value{
a named list of log fold changes with corresponding gene symbols.
}
\description{
Return a list of log fold changes with corresponding gene symbols as their names. This function is calling DESeq function from DESeq2 package, so it takes care of the normalization within the model.
}
