% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNORM.R
\name{getNORM}
\alias{getNORM}
\title{get normalized data}
\usage{
getNORM(countsData)
}
\arguments{
\item{countsData}{a data frame contains all numeric counts. Columns should be the gene symbols and rows should be sample IDs.}
}
\value{
return the same data frame format with normalized counts
}
\description{
This function uses mean-variance method to get the normalized values for a count data. It's calling voom function from limma package.
}
