% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactome.R
\name{getReactome}
\alias{getReactome}
\title{Gets gene sets with Hugo gene symbols to use with SGSEA analysis.}
\usage{
getReactome(species = "human")
}
\arguments{
\item{species}{currently accepts 'human' or 'mouse'}
}
\value{
A list of gene sets from Reactome database
}
\description{
Gets gene sets with Hugo gene symbols to use with SGSEA analysis.
}
