% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSGSEA.R
\name{getSGSEA}
\alias{getSGSEA}
\title{Perform the Survival-based Gene Set Enrichment Analysis}
\usage{
getSGSEA(pathways, stats, minGenes, maxGenes, powerParam = 1)
}
\arguments{
\item{pathways}{A list of pathways annotation (e.g. the result object from getReactome function).}

\item{stats}{A named list of statistics (e.g. the result object from getLFC or getLHR) with corresponding unique gene symbols. The list should not contain any NA's.}

\item{minGenes}{The minimal number of genes that are included in a gene set for the test.}

\item{maxGenes}{The maximum number of genes that are included in a gene set for the test.}

\item{powerParam}{The default value is 1, meaning that gene-level stats are used as it's. If gseaParam is 0, then all gene-level stats effectively become ones
and GSEA test becomes Kolmogorov-Smirnov test. The power of the test that all input gene-level stats are raised to. Also see gseaParam in fgsea function from fgsea package.}
}
\value{
A table of pathway enrichment analysis results. Rows are the tested pathways and columns are the p-values, adjusted p-values, log 2 of the standard errors, enrichment scores,
normalized enrichment scores, sample size for each pathway, and leading edge genes. See value in the fgsea function from fgsea package for more details.
}
\description{
Perform a gene set enrichment analysis based on the survival statistics, log hazard ratios. This function is calling the fgsea function from the fgsea package.
}
