% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIRC.R
\docType{data}
\name{KIRC}
\alias{KIRC}
\title{Kidney Renal clear cell carcinoma(KIRC)}
\format{
\subsection{\code{KIRC}}{

A data frame with 520 patient IDs (rows) and 3000 gene symbols (columns) :
\describe{
\item{ID}{patient's ID}
\item{survtime}{patient's survival time in days}
\item{status}{1 for alive, 2 for dead}
\item{barcode}{patient's barcode}
...
}
}
}
\source{
\url{https://gdac.broadinstitute.org/#}
}
\usage{
KIRC
}
\description{
A subset of Kidney Renal Clear Cell Carcinoma mRNA seq data (version 2016_01_28) from Broad Institute The Cancer Genome Atlas (TCGA) Genome Data Analysis Center (GDAC) Firehose website.
}
\keyword{datasets}
