% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIRCRNAseq.R
\docType{data}
\name{KIRCRNAseq}
\alias{KIRCRNAseq}
\title{Kidney Renal clear cell carcinoma(KIRC)}
\format{
\subsection{\code{KIRCRNAseq}}{

A data frame with 520 rows for patient IDs and 3000 columns for gene symbols:
\describe{
\item{ID}{patient's ID}
\item{survtime}{patient's survival time in days}
\item{status}{1 for alive, 2 for dead}
\item{barcode}{patient's barcode}
...
}
}
}
\source{
\url{http://firebrowse.org/#}
}
\usage{
KIRCRNAseq
}
\description{
A subset of data from Broad Institute of MIT & Harvard FireBrowse
}
\keyword{datasets}
