% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Launch to a corresponding SGSEA Shiny app}
\usage{
runExample()
}
\description{
Need to call the package name first to avoid conflict functions from other packages (e.g. SGSEA::runExample())
}
