% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGSEA-package.R
\docType{package}
\name{SGSEA-package}
\alias{SGSEA}
\alias{SGSEA-package}
\title{SGSEA: Survival-based Gene Set Enrichment Analysis}
\description{
Perform a survival-based gene set enrichment analysis, the log hazard ratio, by calling the functions from survival and fgsea package, and compare with pathways from Reactome database. Function for extracting human pathways from Reactome database is included. The package also provides mean-variance model based normalization method from limma package and log fold change calculation from DESeq2 package. The package also includes a data set named KIRC for users to play with it. This data is a subset of Kidney Renal Clear Cell Carcinoma mRNA seq data (520 patients with 20502 genes, version 2016_01_28) from the Broad institute The Cancer Genome Atlas (TCGA) Genome Data Analysis Center (GDAC) Firehose website (http://gdac.broadinstitute.org).
}
\author{
\strong{Maintainer}: Xiaoxu Deng \email{xdeng3@kumc.edu}

Authors:
\itemize{
  \item Jeffrey Thompson \email{jthompson21@kumc.edu}
}

}
\keyword{internal}
