% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichPlot.R
\name{getEnrichPlot}
\alias{getEnrichPlot}
\title{Create an individual enrichment plot for a specific pathway}
\usage{
getEnrichPlot(pathways, pathwayName, stats)
}
\arguments{
\item{pathways}{A list of human pathways imported from the Reactome database. Same object as the argument in getSGSEA function.}

\item{pathwayName}{Enter the whole name of a specific pathway from getSGSEA result table with the quotation mark.}

\item{stats}{Log hazard ratios or log fold changes.}
}
\value{
An enrichment plot for a specific pathway
}
\description{
Create an individual enrichment plot for a specific pathway
}
