% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLHR.R
\name{getLHR}
\alias{getLHR}
\title{Generate a named numeric vector of the log hazard ratios from Cox Proportional Hazards Model}
\usage{
getLHR(normalizedData, survTime, survStatus)
}
\arguments{
\item{normalizedData}{A data frame contains the normalized counts. Rows must be the sample IDs. Columns must be the unique gene symbols. See KIRC data included in this package as the input example format.}

\item{survTime}{A numeric vector of patient's survival times. User should be aware of matching the gene symbols from the normalized count data with each individual's survival times.}

\item{survStatus}{A numeric vector of patient's survival status. Normally 0=alive, 1=dead. Or 1=alive, 2=death. If omitted, all subjects are assumed to have an event. User should be aware of matching the gene symbols from the normalized count data with each individual's survival status.}
}
\value{
A named list of the log hazard ratios with their corresponding gene symbols.
}
\description{
Return a numeric vector of log hazard ratios with corresponding gene symbols as their names. This function uses the coxph function from survival package. Data should be filtered and normalized before entering this step.
}
