% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLHR.R
\name{getLHR}
\alias{getLHR}
\title{Generate log hazard ratios from Cox Proportional Hazards Model}
\usage{
getLHR(normalizedData, survTime, survStatus)
}
\arguments{
\item{normalizedData}{a data frame contains normalized counts. Columns should be the gene symbols and rows should be sample IDs.
Data should also be filtered before entering this step.}

\item{survTime}{a numeric vector of survival times. User should be aware of keeping the same order of gene symbols for survival times and the genes in normalizedData.}

\item{survStatus}{a numeric vector indicating events. Normally 0=alive, 1=dead. Or 1=alive, 2=death. If omitted, all subjects are assumed to have an event.
User should be aware of keeping the same order of gene symbols for survival status and the genes in normalizedData.}
}
\value{
a named list of log hazard ratios with corresponding gene symbols.
}
\description{
Return a list of log hazard ratios with corresponding gene symbols as their names. This function is calling coxph function from survival package.
}
