% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorm.R
\name{getNorm}
\alias{getNorm}
\title{Normalized the raw counts data}
\usage{
getNorm(countsData, plot = FALSE)
}
\arguments{
\item{countsData}{A data frame contains all numeric counts.
Rows must be the sample IDs. Columns must be the unique gene symbols.
See the included 'KIRC' dataset for an example.}

\item{plot}{Logical. If \code{TRUE}, a mean-variance plot will be generated. Default is \code{FALSE}.}
}
\value{
a data frame with normalized counts.
}
\description{
This function uses mean-variance method to get a normalized count data.
using the voom function from limma package. Data should be filtered
before entering this step.
}
