% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorm.R
\name{getNorm}
\alias{getNorm}
\title{Normalized the raw counts data}
\usage{
getNorm(countsData)
}
\arguments{
\item{countsData}{A data frame contains all numeric counts. Rows must be the sample IDs. Columns must be the unique gene symbols. See KIRC data included in this package as the input example format.}
}
\value{
Return a data frame with normalized counts.
}
\description{
This function uses mean-variance method to get a normalized count data. It's calling voom function from the limma package. Data should be filtered before entering this step.
}
