% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactome.R
\name{getReactome}
\alias{getReactome}
\title{Extract the gene annotation with Hugo gene symbols from the Reactome database.}
\usage{
getReactome(species = "human")
}
\arguments{
\item{species}{Currently only accepts 'human' or 'mouse'}
}
\value{
A large list of pathway annotation with gene symbols.
}
\description{
Extract the gene annotation with Hugo gene symbols from the Reactome database.
}
