% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSGSEA.R
\name{getSGSEA}
\alias{getSGSEA}
\title{get survival-based gene set enrichment anaysis}
\usage{
getSGSEA(pathways, stats, minGenes, maxGenes, powerParam = 1)
}
\arguments{
\item{pathways}{a list of human pathways imported from Reactome database.}

\item{stats}{a named list of statistics(logHazardRatio or logFoldChange) with corresponding gene symbols. The gene symbols should be unique. The list should not contain any NA's.}

\item{minGenes}{the minimal number of genes that are included in a gene set for the test.}

\item{maxGenes}{the maximum number of genes that are included in a gene set for the test.}

\item{powerParam}{The default value is 1, meaning that gene-level stats are used as it's. If gseaParam is 0, then all gene-level stats effectively become ones
and GSEA test becomes Kolmogorov-Smirnov test. The power of the test that all input gene-level stats are raised to. Also see gseaParam in fgsea function from fgsea package.}
}
\value{
A table of GSEA results.Rows are the tested pathways and columns are
name of the pathway, p-value of the enrichment test, a BH-adjusted p-value, Enrichment Scores,
Normalized enrichment score to the corresponding the sample size, and leading edge. Also see
Value in the fgsea function from fgsea package.
}
\description{
Perform a gene set enrichment analysis based on the survival statistics,
the log hazard ratio, by calling the functions from fgsea package,
and compare with pathways from Reactome database.
}
