% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10.R
\name{getTop10}
\alias{getTop10}
\title{get top 10 and bottom 10 significant pathways}
\usage{
getTop10(sgsea.result, pathways, stats, plotParam)
}
\arguments{
\item{sgsea.result}{a table from getSGSEA results.}

\item{pathways}{the human pathways imported from Reactome. Same object as input in getSGSEA function.}

\item{stats}{a named list of log hazard ratios. Same object as input in getSGSEA function.}

\item{plotParam}{The parameter to adjust the displayed values. If close to 0, will get a flatten plot. The Default is 1.}
}
\value{
a table of top10 and bottom10 significant pathways in a ggplot object with enrichment barcode plots.
}
\description{
get top 10 and bottom 10 significant pathways
}
