% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop10.R
\name{getTop10}
\alias{getTop10}
\title{Generate a table of the top 10 significant pathways with positive NES as well as the top 10 significant pathways with negative NES}
\usage{
getTop10(sgsea.result, pathways, stats, plotParam)
}
\arguments{
\item{sgsea.result}{A data frame (e.g. an object from getSGSEA results).}

\item{pathways}{A list of pathways annotation (e.g. the result object from getReactome function).}

\item{stats}{A named list of statistics (e.g. the result object from getLFC or getLHR) with corresponding unique gene symbols. The list should not contain any NA's.}

\item{plotParam}{The parameter to adjust the displayed values. If close to 0, will get a flatten plot. The Default is 1.}
}
\value{
A data frame of the top 10 significant pathways with positive NES as well as the top 10 significant pathways with negative NES in a ggplot object embedded with the enrichment barcode plots.
}
\description{
Generate a table of the top 10 significant pathways with positive NES as well as the top 10 significant pathways with negative NES
}
