% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KIRC}
\alias{KIRC}
\title{Example KIRC (Kidney Renal clear cell carcinoma) dataset}
\format{
A data frame with 50 patients (rows) and 623 variables (columns):
\describe{
\item{ID}{Patient identifier}
\item{survtime}{Survival time in days}
\item{status}{Survival status (1 = alive, 2 = dead)}
\item{A1BG, A1CF, ..., A4GALT}{Gene expression values for selected genes}
}
}
\source{
\url{https://gdac.broadinstitute.org/#}
}
\usage{
KIRC
}
\description{
A curated subset of Kidney Renal Clear Cell Carcinoma (KIRC) mRNA-seq data (version 2016_01_28) from the TCGA GDAC Firehose portal.
This dataset is used to demonstrate the SGSEA workflow.
}
\examples{
data(KIRC)  # lazy loading.
head(KIRC)
}
\keyword{datasets}
