% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGSEA-package.R
\docType{package}
\name{SGSEA-package}
\alias{SGSEA}
\alias{SGSEA-package}
\title{SGSEA: Survival-based Gene Set Enrichment Analysis}
\description{
Perform a Survival-based Gene Set Enrichment Analysis (SGSEA) and the typical case vs. control (or normal vs. tumor) GSEA by calling functions from the survival and fgsea package. The gene annotation database included in this package is the Reactome database. The package also provides a mean-variance model based normalization method for raw count data which is from the limma package. The log fold change calculation function is calling from DESeq2 package, so the user does not need to normaliza the raw count data when conducting the typical GSEA method. The package also includes a data set named KIRC which gives an example of what an appropriate input data format should be. This data is a subset of Kidney Renal Clear Cell Carcinoma mRNA seq data (originally 520 patients with 20502 genes, version 2016_01_28) from the Broad institute The Cancer Genome Atlas (TCGA) Genome Data Analysis Center (GDAC) Firehose website (http://gdac.broadinstitute.org).
}
\author{
\strong{Maintainer}: Xiaoxu Deng \email{xdeng3@kumc.edu}

Authors:
\itemize{
  \item Jeffrey Thompson \email{jthompson21@kumc.edu}
}

}
\keyword{internal}
